Changed Cache system
Added ability to save images to cache. Added BufferedImageBlock, and changed cache to handle just Block rather than specific types of Block. This means whatever is sending/receiving blocks to/from cache should cast the Block to the type they are expecting to be able to get the data back out.